Reorganize and deduplicate label widgets
